disable console reading text
the thread that reads messages from the console is buggy and not used. this code actually slows down Splashy and causes SEGV and ABRT signals to be caught when exiting with ESC while in Splashy "test" mode. We need to revisit this feature later. splashy_update "text foo" should still work and allow users to send messages to the text overlay. F2 also works