makes splashy_update static
even though splashy_update is really use after the system is already running, sometimes we need to use this command very early or over the network (LTSP) and it needs to have all its libraries included in itself. For now, we are shipping this static as well as Splashy binary itself. Later we will drop all static