code cleanups. xres/yres implies resolution
ok, this makes a distinction between the surface/screen/window all treated as the same thing. width and height really means the size of one of those 3 (in fullscreen mode, it makes no difference). We now use xres and yres to refer to the resolution of the actual screen as reported by /dev/fb* (loaded framebuffer). minor and subtle way of thinking about it