do not trust the resolution as reported by primary_{window,}_surface->GetSize()
the resolution reported by the ->GetSize() call for the *surface tends to be according to whatever size is reported by directfb from adding (double+) the size of the actual screen size. This means that on dual-head video cards, the height is twice as much as it should be (from what ioctl() reports from /dev/fb0). Since Splashy is only concern on the size of the screen visible by the user, and not the view that it's possible to go to, we introduced a _get_screen_size() helper that should eliminate this problem